Add files via upload
This is the script for snps_subset dataset that gives the box plot, histogram, and the upset plot. The upset plot is optional, I just tried to keep it to show the traits distribution for each specie or we can say the degree of species variation. But it didn't turn out like what I expected. So, if we really want to keep that plot for snps_subset dataset we will make the changes later.